+++ view-contacts/autotests/autotests-happy-path.sh
@@ -1,14 +1,16 @@
 #!/usr/bin/env bash
 
 # Test setup
+APP_HANDLER="./tteams-contacts-handler"
+APP="./tteams-contacts"
 HANDLER_STDIN=handler-stdin
 HANDLER_STDOUT=handler-stdout
 SHARED_NAME=contacts
 mkfifo ${HANDLER_STDIN}
 sleep infinity > ${HANDLER_STDIN} &
 HANDLER_STDIN_PID=$!
-./tteams-contacts-handler "${SHARED_NAME}" < "${HANDLER_STDIN}" &
-./tteams-contacts 0 0 "${SHARED_NAME}" &> "${HANDLER_STDOUT}" &
+${APP_HANDLER} "${SHARED_NAME}" < "${HANDLER_STDIN}" &
+${APP} 0 0 "${SHARED_NAME}" &> "${HANDLER_STDOUT}" &
 
 # Test scenario
 sleep 3 # Wait for synchronization
@@ -25,31 +27,44 @@ ACTUAL_RESULTS=$(<"${HANDLER_STDOUT}")
 
 # Test cleanup
 kill $HANDLER_STDIN_PID
-wait $(pidof ./tteams-contacts-handler)
-wait $(pidof ./tteams-contacts)
-rm -f ${HANDLER_STDIN}
-rm -f ${HANDLER_STDOUT}
+sleep 3 # Wait for stop
 
 # Test verdict
+EXIT_STATUS=0
+APP_PID=$(pgrep -f ${APP} | head -n 1)
+if [[ "${APP_PID}" ]]; then
+    echo "Error: Application is still running! Killing pid=$APP_PID..."
+    kill $APP_PID
+    EXIT_STATUS=1
+fi
+APP_HANDLER_PID=$(pgrep -f ${APP_HANDLER} | head -n 1)
+if [[ "${APP_HANDLER_PID}" ]]; then
+    echo "Error: Application handler is still running! Killing pid=$APP_HANDLER_PID..."
+    kill $APP_HANDLER_PID
+    EXIT_STATUS=1
+fi
 if [ -f "/dev/shm/${SHARED_NAME}" ]; then
     echo "Error: File /dev/shm/${SHARED_NAME} exists!"
-    exit 1
+    EXIT_STATUS=1
 fi
 if [ -f "/dev/shm/sem.${SHARED_NAME}-data-consumed" ]; then
     echo "Error: File /dev/shm/sem.${SHARED_NAME}-data-consumed exists!"
-    exit 1
+    EXIT_STATUS=1
 fi
 if [ -f "/dev/shm/sem.${SHARED_NAME}-data-produced" ]; then
     echo "Error: File /dev/shm/sem.${SHARED_NAME}-data-produced exists!"
-    exit 1
+    EXIT_STATUS=1
 fi
-if [[ "$ACTUAL_RESULTS" == "$EXPECTED_RESULTS" ]]; then
-    echo "Success: Test passed!"
-    exit 0
-else
+if [[ "$ACTUAL_RESULTS" != "$EXPECTED_RESULTS" ]]; then
     echo "Error: Test failed!"
     printf "%s" "$ACTUAL_RESULTS" > actual_results.txt
     printf "%s" "$EXPECTED_RESULTS" > expected_results.txt
-    exit 1
+    EXIT_STATUS=1
 fi
 
+if [[ $EXIT_STATUS -eq 0 ]]; then
+    echo "Success: Test passed!"
+    rm -f ${HANDLER_STDIN}
+    rm -f ${HANDLER_STDOUT}
+fi
+exit $EXIT_STATUS
